+++ BeaconSdk/MatrixSdk/Domain/MatrixRoomFactory.cs
@@ -1,9 +1,8 @@
-namespace MatrixSdk.Application
+namespace MatrixSdk.Domain
 {
     using System.Collections.Generic;
-    using Domain;
-    using Domain.Room;
     using Infrastructure.Dto.Sync;
+    using Room;
 
     public class MatrixRoomFactory
     {
